+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## v3.4.0
+
+### Added
+
+- Make `Coordinate::fromString()` more flexible to ignore leading 0 in column numbers, e.g. A01
+
 ## v3.3.0
 
 ### Added